清理cleancode
# message auto-generated for no-merge-commit merge:
merge cherry-pick-mr-39-1740399236958-auto into master

清理cleancode


Match-id-c0b1c97145c5e236af2b6a12593caabbd2e9983b

+++ src/cluster/client/cm_client_event.c
@@ -168,7 +168,7 @@ int32_t CM_SetPtFinishStatus(uint16_t poolId, uint16_t ptNum, PtFinish *eventLis
         return ret;
     }
 
-    uint16_t len = sizeof(CmNodeEvent) + sizeof(PtFinish) * ptNum;
+    size_t len = sizeof(CmNodeEvent) + sizeof(PtFinish) * ptNum;
     if (len > NODE_META_BUFF_LEN) {
         CM_LOGERROR("PtNum(%u) too much, poolId(%u).", ptNum, poolId);
         return CM_ERR;
